Added support for Mongoose
I've saw
[`relay-mongodb-connection`](https://github.com/mikberg/relay-mongodb-connection),
loved it and wanted to do the same for mongoose users.
Most of the code is basically the same, just few changes to match
mongoose API.

Added the function `connectionFromMongooseQuery` and
`connectionFromMongooseAggregate` to create a connection from mongoose
`Query` or `Aggregate` respectively.

Didn't want to open an individual package, since most of the code is the
same and NOT MINE.

Haven't done much testing, just enough to cover all my use cases.